+++ test/bcld_test.sh
@@ -110,6 +110,20 @@ function BCLD_KEYMAPs () {
     list_param "${setxkbmap_version}" 'setxkbmap version'
 }
 
+## Function to display if BCLD power settings are enabled
+function BCLD_POWER () {
+
+        # Check client shutdown timer
+		if [[ -n "${BCLD_SHUTDOWN}" ]]; then
+			list_param "${BCLD_SHUTDOWN}" 'BCLD client shutdown timer'
+		fi
+
+        # Check client restart timer
+		if [[ -n "${BCLD_RESTART}" ]]; then
+			list_param "${BCLD_RESTART}" 'BCLD restart shutdown timer'
+		fi
+}
+
 ## Function to display battery in TEST console
 function BCLD_BAT () {
         # Check laptop battery, if present
@@ -435,6 +449,7 @@ function reset_terminal () {
     list_param "${BCLD_SPEED}" 'Link speed (Megabytes/s)'
     list_param "${PACKET_LOSS}" 'Packets dropped (so far)'
     BCLD_KEYMAPs
+    BCLD_POWER
     BCLD_BAT
     list_exit